get FindDaemonDialog working
Removal of `Gtk.Dialog.run` is quite problematic given how
`FindDaemonDialog` is invoked, but rewriting the dialog through 2 layers
of shared inheritance to obey the newer async signal-based pattern will
be a huge change.

`EditorDialog` already has a reference to `app`, so we can get away with
this hack for now... but I'd definitely like to see it refactored in the
future.